Uncapped framerate

+++ i_system.pas
@@ -17,7 +17,7 @@
 //
 //  You should have received a copy of the GNU General Public License
 //  along with this program; if not, write to the Free Software
-//  Foundation, inc., 59 Temple Place - Suite 330, Boston, MA
+//  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA
 //  02111-1307, USA.
 
 //------------------------------------------------------------------------------
@@ -37,7 +37,7 @@ procedure I_Init;
 { Called by D_DoomLoop, }
 { returns current time in tics. }
 function I_GetTime: integer;
-function I_GetTime64: integer;
+function I_GetTime64: int64;
 
 procedure I_ClearInterface(var Dest: IInterface);
 
@@ -163,7 +163,7 @@ begin
   Result := round((_time - basetime) * invFreq);
 end;
 
-function I_GetTime64: integer;
+function I_GetTime64: int64;
 var
   _time: int64;
 begin
